fix recording artist field type

+++ database/migrations/2021_02_12_133858_create_recordings_table.php
@@ -16,7 +16,7 @@ class CreateRecordingsTable extends Migration
         Schema::create('recordings', function (Blueprint $table) {
             $table->id();
             $table->string('title');
-            $table->integer('artist');
+            $table->unsignedBigInteger('artist');
             $table->foreign('artist')->references('id')->on('artists');
             $table->integer('length');
             $table->string('ISRC');
